Pass through metadata to be used for statistics rendering; fix issues with hover-card
